moves constrains and their validators to more appropriate namespace
